refactor(measurement): split benchmark inputs and planning
